requirements.txt
# Core data science - specify versions compatible with ARM64
numpy>=1.26.4,<1.26.4
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
jupyter>=1.0.0
notebook>=6.4.0

# Machine Learning
scikit-learn>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
lightgbm>=3.3.0
xgboost>=1.5.0

# Security & Malware Analysis
lief>=0.13.0
pefile>=2019.4.0
capstone>=4.0.0

# Additional utilities
seaborn>=0.17.0
tqdm>=4.62.0
joblib>=1.1.0
pillow>=8.3.0

# EMbER specific if needed
ember-features>=0.1.0
